Add LiteLLM backend with watsonx (and ollama) support
This might not replace OpenAI backend because it has limited
support, but it does know how to work with watsonx (and ollama
and some others).

We use aconfig, but clients typically honor environment variables
to set endpoints, model names, api keys, etc...  So this is kind
of a mix.  The example demonstrates how to use a .env to switch
between backends.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>