feat: add example for how to use Replicate
We can use LiteLLM to access Replicate

The examples/watsonx_litellm.py explains how to use environment variables
to make LiteLLM work with watsonx, Ollama, and now Replicate (with a
thinking/non-thinking example).

This example filename is not great, but emphasizing watsonx is OK.
Need to follow-up with better docs explaining what is supported and
link to examples from there. Separate issue coming for that.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>
Signed-off-by: Louis Mandel <lmandel@us.ibm.com>