bug: remove httpx dependency and timeout field
To support timeout for openai and litellm, I introduced an httpx
dependency that we don't want in io.  We can remove timeout
field since we don't add any value with it (would only limit
it's httpx features).

This should fall into "other params" that we don't want to have
an opinion about.  If/when we do care, we'll need to handle httpx
in/out for this dataclass in an import optional way.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>
Signed-off-by: Louis Mandel <lmandel@us.ibm.com>