Fix unit tests
Some backends like LiteLLM require stop attribute to be an array of strings instead of just
a string. This fix updates string to an array if required.

The LitleLLM stop cassette is re-recorded with the fix and which produces the expected output in this instance.
As a consequence of the correct cassette, the offending unit test now passes.

Thanks Mark Sturdevant for finding the fix.

Co-authored-by: Mark Sturdevant <mark.sturdevant@ibm.com>
Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>